# _examples/petStore/openapi.gen.yaml
#  vTODO 5b85c752b441f69fd2f86fa3f69d9be865033905
# --
# Code generated by webrpc-gen@v0.10.x-dev with openapi@v0.10.0 generator; DO NOT EDIT
# 
# gospeak ./schema/api.go golang@v0.10.0 -server -pkg=server -out=./server/server.gen.go golang@v0.10.0 -client -pkg=client -out=./client/client.gen.go typescript@v0.10.0 -client -out=./client.gen.ts openapi@v0.10.0 -out=./openapi.gen.yaml
openapi: 3.0.0
info:
  title: 'PetStore API'
  version: ''
servers:
  - url: ''
    description: ''
components:
  schemas:
    empty:
      type: object
    error:
      properties:
        error:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
    Pet:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        available:
          type: boolean
        photoUrlS:
          type: string
        tags:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/tag'
    CreatePet_Request:
      properties:
        new:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Pet'
    DeletePet_Request:
      properties:
        ID:
          type: number
    GetPet_Request:
      properties:
        ID:
          type: number
    ListPets_Request:
      type: object
    UpdatePet_Request:
      properties:
        ID:
          type: number
        update:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Pet'
    CreatePet_Response:
      properties:
        pet:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Pet'
    DeletePet_Response:
      type: object
    GetPet_Response:
      properties:
        pet:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Pet'
    ListPets_Response:
      properties:
        pets:
          type: array
          description: '[]Pet'
          items:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/Pet'
    UpdatePet_Response:
      properties:
        pet:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Pet'

paths:
  /rpc/PetStore/CreatePet:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePet_Request'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePet_Response'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /rpc/PetStore/DeletePet:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeletePet_Request'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletePet_Response'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /rpc/PetStore/GetPet:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPet_Request'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPet_Response'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /rpc/PetStore/ListPets:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListPets_Request'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPets_Response'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /rpc/PetStore/UpdatePet:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePet_Request'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePet_Response'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'